refactor: remove floating First Mate cockpit + bottom notification card
Delete the centered floating command palette (HelmCockpitController) and the
transient bottom-anchored notification (HelmFloatingCard), plus the cockpit-only
radar orb (HelmOrbView). The command composer already lives in the dashboard
overview, so all entry points repoint there:

- space no longer opens the cockpit; `?` help + Esc-dismiss rehomed onto
  DashboardViewController (toggleHelp/closeTopmostOverlay), reusing the kept
  KeyboardHelpOverlay
- MenuRowButton extracted to its own file (overview composer still uses it)
- First Mate menu commands + /new route through startNewCommand(prefill:)
- HelmCommandOutcome moved onto MainWindowController; radar wiring dropped
